requirements.txt
numpy
pandas
pyexpect
openai
python-dotenv
discord
pillow
wikipedia
IPython
tiktoken
pyowm
duckduckgo-search
youtube-transcript-api
requests
beautifulsoup4
pyyaml
PySelenium
py-webdriver-manager
google-api-python-client
geopy